Refactor into one single endpoint
