+++ front_end/src/Components/EditPayroll/api.js
@@ -2,16 +2,12 @@ import { getData, postData } from "../../Util";
 
 import * as types from "./types";
 
-const vacanciesSlug = "vacancies/";
-const payModifiersSlug = "pay_modifiers/";
-const previousMonthsSlug = "previous_months/";
-
 /**
  * Fetch payroll data and return it as a promise.
  * @returns {Promise<types.PayrollData[]>} A promise resolving to an array of objects containing employee information.
  */
 export function getPayrollData() {
-  return getData(getPayrollApiUrl()).then((data) => data.data);
+  return getData(getPayrollApiUrl());
 }
 
 /**
@@ -22,60 +18,6 @@ export function getPayrollData() {
  */
 export function postPayrollData(payrollData) {
   return postData(getPayrollApiUrl(), JSON.stringify(payrollData));
-}
-
-/**
- * Fetch vacancy data and return it as a promise.
- * @returns {Promise<types.VacancyData[]>} A promise resolving to an array of objects containing vacancy information.
- */
-export function getVacancyData() {
-  return getData(getPayrollApiUrl() + vacanciesSlug).then((data) => data.data);
-}
-
-/**
- * Post modified vacancy data.
- *
- * @param {types.VacancyData[]} vacancyData - Vacancy data to be sent.
- * @returns {import("../../Util").PostDataResponse} Updated vacancy data received.
- */
-export function postVacancyData(vacancyData) {
-  return postData(
-    getPayrollApiUrl() + vacanciesSlug,
-    JSON.stringify(vacancyData),
-  );
-}
-
-/**
- * Fetch pay modifier data and return it as a promise.
- * @returns {Promise<types.PayModifierData[]>} A promise resolving to an array of objects containing pay modifier information.
- */
-export function getPayModifierData() {
-  return getData(getPayrollApiUrl() + payModifiersSlug).then(
-    (data) => data.data,
-  );
-}
-
-/**
- * Post modified pay modifiers data.
- *
- * @param {types.PayModifierData[]} payModifierData - Pay modifier data to be sent.
- * @returns {import("../../Util").PostDataResponse} Updated pay modifier data received.
- */
-export function postPayModifierData(payModifierData) {
-  return postData(
-    getPayrollApiUrl() + payModifiersSlug,
-    JSON.stringify(payModifierData),
-  );
-}
-
-/**
- * Fetch previous months data and return it as a promise.
- * @returns {Promise<types.PreviousMonthsData[]>} A promise resolving to an array of objects containing previous months information.
- */
-export function getPreviousMonthsData() {
-  return getData(getPayrollApiUrl() + previousMonthsSlug).then(
-    (data) => data.data,
-  );
 }
 
 /**
